+++ src/combinator/apply-self.ts
@@ -1,8 +1,7 @@
-import { $, Type, Kind } from "..";
-import { RecursiveKind } from "./recursive-kind";
+import { $, Type, Kind, Combinator } from "..";
 
 export abstract class ApplySelf extends Kind.Kind {
   abstract f: (
-    x: Type._$cast<this[Kind._], RecursiveKind>
+    x: Type._$cast<this[Kind._], Combinator.RecursiveKind>
   ) => $<typeof x, Type._$cast<typeof x, Kind._$inputOf<typeof x>>>;
 }